Re-organized classes
